+++ frontend/app/src/components/settings/general/AbbreviateNumberSetting.vue
@@ -1,0 +1,33 @@
+<template>
+  <settings-option
+    #default="{ error, success, update }"
+    setting="abbreviateNumber"
+    frontend-setting
+  >
+    <v-switch
+      v-model="abbreviate"
+      data-cy="frontend-settings__fields__abbreviate_number"
+      :label="tc('frontend_settings.label.abbreviate_number')"
+      :hint="tc('frontend_settings.subtitle.abbreviate_number')"
+      color="primary"
+      persistent-hint
+      :success-messages="success"
+      :error-messages="error"
+      @change="update"
+    />
+  </settings-option>
+</template>
+
+<script setup lang="ts">
+import { Ref } from 'vue';
+import { useFrontendSettingsStore } from '@/store/settings/frontend';
+
+const abbreviate: Ref<boolean> = ref(false);
+const { abbreviateNumber } = storeToRefs(useFrontendSettingsStore());
+
+onMounted(() => {
+  set(abbreviate, get(abbreviateNumber));
+});
+
+const { tc } = useI18n();
+</script>